common: Edit flags supported changes to IdentityInfo.
This commit adds 'EditFlags' to IdentityInfo. Which helps to track
the changes happening to a given identity info. This field is used
especially in DB to save only changed fields, than blindly refreshing
whole identity info.

I would prapose to change the IdentityInfo type to a new Boxed type
which internally uses the GSignondDictionary to save data. So that we
can move the 'EditFlags' out of the dictionary.